requirements.txt
python-telegram-bot==20.8
geopy
stripe
fpdf
aiohttp
pillow
